Fix change to doubles

+++ src/functions/Derivatives.java
@@ -1,5 +1,5 @@
 package functions;
 
 public interface Derivatives {
-    float calculateDerivative(int order, int step);
+    double calculateDerivative(int order, int step);
 }
