+++ src/functions/FunctionWithDerivatives.java
@@ -7,19 +7,19 @@ import java.util.Map;
 import java.util.function.Function;
 
 public class FunctionWithDerivatives implements Derivatives {
-    private final Map<Integer, Map<Integer, Float>> derivativeFunctions;
+    private final Map<Integer, Map<Integer, Double>> derivativeFunctions;
 
-    public FunctionWithDerivatives(Map<Integer, Float> function) {
+    public FunctionWithDerivatives(Map<Integer, Double> function) {
         this.derivativeFunctions = new HashMap<>();
         this.derivativeFunctions.put(0, function);
     }
 
-    public void setDerivative(int order, Map<Integer, Float> derivativeFunction) {
+    public void setDerivative(int order, Map<Integer, Double> derivativeFunction) {
         this.derivativeFunctions.put(order, derivativeFunction);
     }
 
     @Override
-    public float calculateDerivative(int order, int step) {
+    public double calculateDerivative(int order, int step) {
         return derivativeFunctions.get(order).get(step);
     }
